backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Gerson Vieira Pedro",email = "80074869+GersonVieiraPedro@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "azure-storage-file-datalake (>=12.20.0,<13.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
